+++ src/lob_asset_management/controller/market.clj
@@ -1,12 +1,10 @@
 (ns lob-asset-management.controller.market
-  (:require [lob-asset-management.io.http_in :as io.http]
+  (:require [clojure.tools.logging :as log]
+            [lob-asset-management.io.http_in :as io.http]
             [lob-asset-management.io.file-in :as io.f-in]
             [lob-asset-management.io.file-out :as io.f-out]
             [lob-asset-management.adapter.asset :as a.a]
-            [lob-asset-management.aux.time :as aux.t]
-            ;[clj-time.core :as t]
-            [java-time.api :as t]
-            ))
+            [lob-asset-management.aux.time :as aux.t]))
 
 (defn keyword-space->underline [m]
   (zipmap (map #(keyword (clojure.string/replace (name %) " " "_")) (keys m))
@@ -24,11 +22,19 @@
           {}
           price-historic))
 
+(defn market-info->last-refreshed-dt
+  [formatted-data]
+  (when-let [latest-refreshed-dt (-> formatted-data :meta-data :3._Last_Refreshed)]
+    (-> latest-refreshed-dt
+        (clojure.string/split #" ")
+        first
+        keyword)))
+
 (defn last-price
   [formatted-data]
   (if (or (not (nil? formatted-data))
           (not (empty? formatted-data)))
-    (if-let [latest-refreshed-dt (-> formatted-data :meta-data :3._Last_Refreshed keyword)]
+    (if-let [latest-refreshed-dt (market-info->last-refreshed-dt formatted-data)]
       (let [latest-refreshed-price (-> formatted-data
                                        :time-series
                                        latest-refreshed-dt
@@ -42,8 +48,8 @@
        :date       latest-refreshed-dt
        :updated-at (aux.t/get-current-millis)
        :historic   price-historic})
-      (throw (ex-info "[ERROR] (2) Something was wrong in get market data => formatted-data" formatted-data)))
-    (throw (ex-info "[ERROR] (2) Something was wrong in get market data => formatted-data" formatted-data))))
+      (throw (ex-info "[last-price] (1) Something was wrong in get market data => formatted-data" formatted-data)))
+    (throw (ex-info "[last-price] (2) Something was wrong in get market data => formatted-data" formatted-data))))
 
 (defn get-b3-market-price
   [asset]
@@ -51,7 +57,7 @@
     (let [formatted-data (formatted-data market-info)
           last-price (last-price formatted-data)]
       last-price)
-    (println "[ERROR] Something was wrong in get market data")))
+    (log/error "[get-b3-market-price] Something was wrong in get market data")))
 
 (defn in-ticket->out-ticket
   [{:asset/keys [ticket type]}]
@@ -65,8 +71,7 @@
    asset-id
    {:keys [price updated-at date historic]}]
   (if (= id asset-id)
-    (let [updated-historic (first (conj (:asset.market-price/historic asset)
-                                        historic))]
+    (let [updated-historic (merge (:asset.market-price/historic asset) historic)]
       (assoc asset :asset.market-price/price price
                    :asset.market-price/updated-at updated-at
                    :asset.market-price/price-date date
@@ -85,28 +90,28 @@
   ([]
    (if-let [assets (io.f-in/get-file-by-entity :asset)]
      (update-asset-market-price assets)
-     (println "[ERROR] update-asset-market-price - can't get assets")))
+     (log/error "[MARKET-UPDATING] update-asset-market-price - can't get assets")))
   ([assets]
    (if-let [less-updated-asset (-> assets (a.a/get-less-market-price-updated 1 1) first)]
-     (do (println "[MARKET-UPDATING] Stating get asset price for " less-updated-asset)
+     (do (log/info "[MARKET-UPDATING] Stating get asset price for " (:asset/ticket less-updated-asset))
          (let [less-updated-asset-ticket (in-ticket->out-ticket less-updated-asset)
                market-last-price (get-b3-market-price less-updated-asset-ticket)
                updated-assets (update-assets assets less-updated-asset market-last-price)]
-           (println "[MARKET-UPDATING] Success " less-updated-asset-ticket " price " (:price market-last-price))
+           (log/info "[MARKET-UPDATING] Success " less-updated-asset-ticket " price " (:price market-last-price))
            (io.f-out/upsert updated-assets)))
-     (println "[WARNING] No asset to be updated"))))
+     (log/warn "[MARKET-UPDATING] No asset to be updated"))))
 
 (defn get-asset-market-price
   "Receive a list of assets and return the list updated without read or write data"
   [assets]
   (if-let [less-updated-asset (-> assets (a.a/get-less-market-price-updated) first)]
-    (do (println "[MARKET-PRICE] Stating get asset price for " less-updated-asset)
+    (do (log/info "[MARKET-PRICE] Stating get asset price for " (:asset/ticket less-updated-asset))
         (let [less-updated-asset-ticket (in-ticket->out-ticket less-updated-asset)
               market-last-price (get-b3-market-price less-updated-asset-ticket)
               updated-assets (update-assets assets less-updated-asset market-last-price)]
-          (println "[MARKET-UPDATING] Success " less-updated-asset-ticket " price " (:price market-last-price))
+          (log/info "[MARKET-UPDATING] Success " less-updated-asset-ticket " price " (:price market-last-price))
           updated-assets))
-    (println "[WARNING] No asset to be updated")))
+    (log/warn "[MARKET-UPDATING] No asset to be updated")))
 
 (comment
   (def aux-market-info (io.http/get-daily-adjusted-prices "CAN"))